+++ src/routes/graphql/types/members.ts
@@ -1,0 +1,4 @@
+export enum MemberId {
+    BASIC = 'basic',
+    BUSINESS = 'business',
+  }
